tests/requirements.txt
psycopg2-binary
pytest
sqlparse
